+++ e2e/testdata/fn-render/generator-absolute-path/.krmignore
@@ -1,0 +1,1 @@
+.expected
