+++ e2e/testdata/fn-render/generator-absolute-path/resources.yaml
