+++ e2e/testdata/fn-render/path-index-current/starlark-httpbin-gen.yaml
